menambahkan kriteria seeder

+++ database/migrations/2023_09_25_151136_create_attributes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama');
             $table->string('kode');
-            $table->integer('bobot');
+            $table->float('bobot');
             $table->timestamps();
         });
     }
